adjust Game Model, Create GameDetail Model and Response. But there's a bug while running getDetailGames

+++ GameCatalog/DetailPopularGame/StarsRatingView.swift
@@ -10,7 +10,7 @@ import UIKit
 
 class StarsRatingView: UIStackView {
     
-    var rating: Double = 0 {
+    var rating: Float = 0 {
         didSet {
             updateStars()
         }
